fix(inference): keep private endpoint proof opaque
